rename stochastic pulse gradient file
