+++ pennylane/gradients/__init__.py
@@ -320,7 +320,7 @@ from . import parameter_shift_hessian
 from . import finite_difference
 from . import spsa_gradient
 from . import hadamard_gradient
-from . import pulse_gradient
+from . import stoch_pulse_gradient
 
 from .gradient_transform import gradient_transform, SUPPORTED_GRADIENT_KWARGS
 from .hessian_transform import hessian_transform
@@ -332,7 +332,7 @@ from .vjp import compute_vjp, batch_vjp, vjp, compute_vjp_multi, compute_vjp_sin
 from .jvp import batch_jvp, jvp, compute_jvp_multi, compute_jvp_single
 from .spsa_gradient import spsa_grad
 from .hadamard_gradient import hadamard_grad
-from .pulse_gradient import stoch_pulse_grad
+from .stoch_pulse_gradient import stoch_pulse_grad
 
 from .hamiltonian_grad import hamiltonian_grad
 from .general_shift_rules import (
