+++ pennylane/gradients/stoch_pulse_gradient.py
